feat: Added table view to test saved and deleting from Core Data
Added an Article entity in Core Data for when we get data from the web
parser. Can now download and saved data to Core Data and added a
temporary table view to test putting data on it, saving it,
and deleting it from core data, all of which works well.